+++ packify/__init__.py
@@ -1,21 +1,5 @@
 from .errors import UsageError
 from .interface import Packable
 from .serialization import pack, unpack, SerializableType
+from .version import __version__, version
 
-
-__version__ = '0.3.2'
-
-
-def version() -> str:
-    """Returns the current version of the packify package."""
-    return __version__
-
-
-__all__ = [
-    'UsageError',
-    'Packable',
-    'pack',
-    'unpack',
-    'SerializableType',
-    'version',
-]